+++ clap/helper.py
@@ -1,0 +1,22 @@
+"""Generates usage information and
+help data from parsers.
+"""
+
+class Helper():
+    """Helper object holds logic used to build usage and
+    help information.
+    """
+    def __init__(self, parser):
+        self.parser = parser
+
+    def usage(self):
+        """Generates usage information.
+        """
+        usage = ''
+        return usage
+
+    def help(self, indent=4):
+        """Generates help information.
+        """
+        info = ''
+        return info